requirements.txt
Flask
gunicorn
numpy
pandas
scikit-learn
tensorflow-cpu